% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.StudentData.R
\name{plot.StudentData}
\alias{plot.StudentData}
\title{Plot Student Performances - Student performance compared to various variables}
\usage{
\method{plot}{StudentData}(data, categorical_var, plot_type = "boxplot")
}
\arguments{
\item{data}{A data-frame containing numeric variables for test scores (e.g. \code{"math.score"}),
and categorical variables that have been converted to factors (e.g. \code{"gender"}).}

\item{categorical_var}{A chosen categorical variable from the dataset (e.g. \code{"gender"}).}

\item{plot_type}{3 different types of plots for analysis (e.g. \code{"boxplot"}).}
}
\value{
A plot, either boxplot, violin or barplot of the \code{"math.score"}, \code{"reading.score"},
or \code{"writing.score"}, compared to the chosen categorical variable (e.g. \code{"gender"}) for analysis.
}
\description{
Plots the subject scores against your chosen categorical variable in 3 ways, "violin", "boxplot" or "barplot".
}
\examples{
data(students)
data <- prepare_student_data(students)
plot.StudentData(data, "gender", "boxplot")



}
