% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.StudentData.R
\name{summary.StudentData}
\alias{summary.StudentData}
\title{Summarize students performances}
\usage{
\method{summary}{StudentData}(obj, group_var, summary_func = mean)
}
\arguments{
\item{obj}{A dataset of class \code{StudentData} containing
at least the variables \code{math.score}, \code{writing.score}, and \code{reading.score}.
Other variables may be used as grouping variables.}

\item{group_var}{A character vector specifying the column name(s) to group the data by.}

\item{summary_func}{A function to summarize each numeric variable
(e.g., \code{mean}, \code{min}, or \code{max}). The function is applied to
\code{math.score}, \code{writing.score}, and \code{reading.score} within each group
defined by \code{group_var}. Defaults to \code{mean}}
}
\value{
A \code{tibble} containing the grouped data and the summarized statistics for
 \code{math.score}, \code{writing.score}, and \code{reading.score}.
}
\description{
This function computes summary statistics (e.g., mean, min, max) for student performance metrics (math.score, writing.score, and reading.score) grouped by specified categorical variables in a dataset.
}
\examples{
data(students)
data <- prepare_student_data(students, scale = FALSE)
summary(obj = data, group_var = c("gender", "lunch"), summary_func = median)
}
